scholl-app/requirements.txt
aioredis==1.3.1
apns2==0.7.1
appdirs==1.4.3
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
autopep8==1.5
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.9
daphne==2.5.0
distlib==0.3.0
distro==1.4.0
Django==2.2.10
django-cors-headers==3.2.1
django-filter==2.2.0
django-polymorphic==2.1.2
django-push-notifications==2.0.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-writable-nested==0.6.0
environs==7.3.0
h2==2.6.2
hiredis==1.0.1
hpack==3.0.0
html5lib==1.0.1
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
lockfile==0.12.2
Markdown==3.2.1
marshmallow==3.5.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.1
pep517==0.7.0
Pillow==7.0.0
progress==1.5
psycopg2==2.8.4
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dotenv==0.12.0
python-monkey-business==1.0.0
pytoml==0.1.21
pytz==2019.3
pywebpush==1.11.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
webencodings==0.5.1
zipp==3.1.0
zope.interface==5.1.0
